refactor(ui): extract skills / techniques / powers / perks tab mixins
Four mid-sized per-tab mixins under l5r/ui/tabs/:

* SkillsTabMixin -> Tab 2 (build_ui_page_2 + on_buy_skill_rank).
* TechniquesTabMixin -> Tab 3 (build_ui_page_3, _build_spell_frame,
  _build_tech_frame, act_memo_spell, act_buy_spell, act_del_spell,
  on_spell_selected).
* PowersTabMixin -> Tab 4 (build_ui_page_4, _build_kata_frame,
  _build_kiho_frame, do_buy_kata, do_buy_kiho).
* PerksTabMixin -> Tab 5 (build_ui_page_5).

main.py: 1204 -> 753 lines.
