+++ l5r/main.py
@@ -62,14 +62,20 @@ from l5r.ui.persistence import PersistenceMixin
 from l5r.ui.tabs.about import AboutTabMixin
 from l5r.ui.tabs.advancements import AdvancementsTabMixin
 from l5r.ui.tabs.pc_info import PcInfoTabMixin
+from l5r.ui.tabs.perks import PerksTabMixin
+from l5r.ui.tabs.powers import PowersTabMixin
 from l5r.ui.tabs.settings_tab import SettingsTabMixin
+from l5r.ui.tabs.skills import SkillsTabMixin
+from l5r.ui.tabs.techniques import TechniquesTabMixin
 from l5r.util import log
 from l5r.util.settings import L5RCMSettings
 
 
 class L5RMain(AboutTabMixin, AdvancementsTabMixin, AdvanceMixin, AdviseMixin,
               HealthDisplayMixin, MenuMixin, NicebarMixin, PcInfoTabMixin,
-              PersistenceMixin, SettingsTabMixin, L5RCMCore):
+              PerksTabMixin, PersistenceMixin, PowersTabMixin,
+              SettingsTabMixin, SkillsTabMixin, TechniquesTabMixin,
+              L5RCMCore):
 
     default_size = QtCore.QSize(820, 720)
     default_point_size = 8.25
@@ -217,377 +223,6 @@ class L5RMain(AboutTabMixin, AdvancementsTabMixin, AdvanceMixin, AdviseMixin,
             vbox.addWidget(grp)
             views_.append(view)
         return mfr, views_
-
-    def _build_spell_frame(self, model, layout):
-        grp = QtWidgets.QGroupBox(self.tr("Spells"), self)
-        hbox = QtWidgets.QHBoxLayout(grp)
-
-        fr_ = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(fr_)
-        vbox.setContentsMargins(3, 3, 3, 3)
-
-        # advantages/disadvantage vertical toolbar
-        def _make_vertical_tb():
-            vtb = widgets.VerticalToolBar(self)
-            vtb.addStretch()
-
-            cb_buy = self.act_buy_spell
-            cb_remove = self.act_del_spell
-            cb_memo = self.act_memo_spell
-
-            self.add_spell_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                self.tr("Add new spell"), cb_buy)
-
-            self.del_spell_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('minus', (16, 16))),
-                self.tr("Remove spell"), cb_remove)
-
-            self.memo_spell_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('book', (16, 16))),
-                self.tr("Memorize/Forget spell"), cb_memo)
-
-            self.del_spell_bt.setEnabled(False)
-
-            vtb.addStretch()
-            return vtb
-
-        # View
-        view = QtWidgets.QTableView(fr_)
-        view.setSizePolicy(QtWidgets.QSizePolicy.Expanding,
-                           QtWidgets.QSizePolicy.Expanding)
-        view.setSortingEnabled(True)
-        view.horizontalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Interactive)
-        view.horizontalHeader().setStretchLastSection(True)
-        view.horizontalHeader().setCascadingSectionResizes(True)
-        view.setModel(model)
-        self.table_views.append(view)
-        sm = view.selectionModel()
-        sm.currentRowChanged.connect(self.on_spell_selected)
-        self.spell_table_view = view
-
-        # Affinity/Deficiency
-        self.lb_affin = QtWidgets.QLabel(self.tr("None"), self)
-        self.lb_defic = QtWidgets.QLabel(self.tr("None"), self)
-
-        aff_fr = QtWidgets.QFrame(self)
-        aff_fr.setSizePolicy(QtWidgets.QSizePolicy.Preferred,
-                             QtWidgets.QSizePolicy.Maximum)
-        fl = QtWidgets.QFormLayout(aff_fr)
-        fl.addRow(self.tr("<b><i>Affinity</i></b>"), self.lb_affin)
-        fl.addRow(self.tr("<b><i>Deficiency</i></b>"), self.lb_defic)
-        fl.setHorizontalSpacing(60)
-        fl.setVerticalSpacing(5)
-        fl.setContentsMargins(0, 0, 0, 0)
-
-        vbox.addWidget(aff_fr)
-        vbox.addWidget(view)
-
-        hbox.addWidget(_make_vertical_tb())
-        hbox.addWidget(fr_)
-        layout.addWidget(grp)
-
-        view.doubleClicked.connect(self.sink4.on_spell_item_activate)
-
-        return view
-
-    def _build_tech_frame(self, model, layout):
-        grp = QtWidgets.QGroupBox(self.tr("Techs"), self)
-        vbox = QtWidgets.QVBoxLayout(grp)
-
-        # View
-        view = QtWidgets.QTableView(self)
-        view.setSortingEnabled(False)
-        view.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.Interactive)
-        view.horizontalHeader().setStretchLastSection(True)
-        view.horizontalHeader().setCascadingSectionResizes(True)
-
-        view.setModel(model)
-        self.table_views.append(view)
-        vbox.addWidget(view)
-        layout.addWidget(grp)
-
-        view.doubleClicked.connect(self.sink4.on_tech_item_activate)
-
-        return view
-
-    def _build_kata_frame(self, model, layout):
-        grp = QtWidgets.QGroupBox(self.tr("Kata"), self)
-        hbox = QtWidgets.QHBoxLayout(grp)
-
-        fr_ = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(fr_)
-        vbox.setContentsMargins(3, 3, 3, 3)
-
-        # advantages/disadvantage vertical toolbar
-        def _make_vertical_tb():
-            vtb = widgets.VerticalToolBar(self)
-            vtb.addStretch()
-
-            cb_buy = self.sink2.act_buy_kata
-            cb_remove = self.sink2.act_del_kata
-
-            self.add_kata_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                self.tr("Add new Kata"), cb_buy)
-
-            self.del_kata_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('minus', (16, 16))),
-                self.tr("Remove Kata"), cb_remove)
-
-            self.add_kata_bt.setEnabled(True)
-            self.del_kata_bt.setEnabled(True)
-
-            vtb.addStretch()
-            return vtb
-
-        # View
-        view = QtWidgets.QTableView(self)
-        view.setSizePolicy(QtWidgets.QSizePolicy.Expanding,
-                           QtWidgets.QSizePolicy.Expanding)
-        view.setSortingEnabled(True)
-        view.horizontalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Interactive)
-        view.horizontalHeader().setStretchLastSection(True)
-        view.horizontalHeader().setCascadingSectionResizes(True)
-        view.setModel(model)
-        view.doubleClicked.connect(self.sink4.on_kata_item_activate)
-        self.ka_table_view = view
-        self.table_views.append(view)
-
-        vbox.addWidget(view)
-
-        hbox.addWidget(_make_vertical_tb())
-        hbox.addWidget(fr_)
-
-        layout.addWidget(grp)
-
-        return view
-
-    def _build_kiho_frame(self, model, layout):
-        grp = QtWidgets.QGroupBox(self.tr("Kiho"), self)
-        hbox = QtWidgets.QHBoxLayout(grp)
-
-        fr_ = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(fr_)
-        vbox.setContentsMargins(3, 3, 3, 3)
-
-        # advantages/disadvantage vertical toolbar
-        def _make_vertical_tb():
-            vtb = widgets.VerticalToolBar(self)
-            vtb.addStretch()
-
-            cb_buy = self.sink2.act_buy_kiho
-            cb_remove = self.sink2.act_del_kiho
-            cb_buy_tattoo = self.sink2.act_buy_tattoo
-
-            self.add_kiho_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                self.tr("Add new Kiho"), cb_buy)
-
-            self.add_tattoo_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                self.tr("Add new Tattoo"), cb_buy_tattoo)
-
-            self.del_kiho_bt = vtb.addButton(
-                QtGui.QIcon(get_icon_path('minus', (16, 16))),
-                self.tr("Remove Kiho"), cb_remove)
-
-            self.add_kiho_bt.setEnabled(True)
-            self.del_kiho_bt.setEnabled(True)
-
-            vtb.addStretch()
-            return vtb
-
-        # View
-        view = QtWidgets.QTableView(self)
-        view.setSizePolicy(QtWidgets.QSizePolicy.Expanding,
-                           QtWidgets.QSizePolicy.Expanding)
-        view.setSortingEnabled(True)
-        view.horizontalHeader().setSectionResizeMode(QtWidgets.QHeaderView.Interactive)
-        view.horizontalHeader().setStretchLastSection(True)
-        view.horizontalHeader().setCascadingSectionResizes(True)
-        view.setModel(model)
-        view.doubleClicked.connect(self.sink4.on_kiho_item_activate)
-        self.ki_table_view = view
-        self.table_views.append(view)
-
-        vbox.addWidget(view)
-
-        hbox.addWidget(_make_vertical_tb())
-        hbox.addWidget(fr_)
-
-        layout.addWidget(grp)
-
-        return view
-
-    def build_ui_page_2(self):
-        self.sk_view_model = models.SkillTableViewModel(self)
-        self.ma_view_model = models.MaViewModel(self)
-
-        # enable sorting through a proxy model
-        self.sk_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.sk_sort_model.setDynamicSortFilter(True)
-        self.sk_sort_model.setSourceModel(self.sk_view_model)
-
-        # enable sorting through a proxy model
-        self.ma_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.ma_sort_model.setDynamicSortFilter(True)
-        self.ma_sort_model.setSourceModel(self.ma_view_model)
-
-        # skills vertical toolbar
-        vtb = widgets.VerticalToolBar(self)
-        vtb.addStretch()
-        vtb.addButton(QtGui.QIcon(get_icon_path('add', (16, 16))),
-                      self.tr("Add skill rank"), self.on_buy_skill_rank)
-        vtb.addButton(QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                      self.tr("Buy skill emphasys"), self.show_buy_emph_dlg)
-        vtb.addButton(QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                      self.tr("Buy another skill"), self.show_buy_skill_dlg)
-        vtb.addStretch()
-
-        models_ = [
-            (
-                "Skills",
-                'table',
-                self.sk_sort_model,
-                None,
-                vtb,
-                self.sink4.on_skill_item_activate
-            ),
-            (
-                self.tr("Mastery Abilities"),
-                'table',
-                self.ma_sort_model,
-                None,
-                None,
-                None
-            )
-        ]
-        frame_, views_ = self._build_generic_page(models_)
-
-        if len(views_) > 0:
-            self.skill_table_view = views_[0]
-
-        self.tabs.addTab(frame_, self.tr("Skills"))
-
-    def build_ui_page_3(self):
-        self.sp_view_model = models.SpellTableViewModel(self)
-        self.th_view_model = models.TechViewModel(self)
-
-        # enable sorting through a proxy model
-        self.sp_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.sp_sort_model.setDynamicSortFilter(True)
-        self.sp_sort_model.setSourceModel(self.sp_view_model)
-
-        frame_ = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(frame_)
-
-        self._build_spell_frame(self.sp_sort_model, vbox)
-        self._build_tech_frame(self.th_view_model, vbox)
-
-        self.tabs.addTab(frame_, self.tr("Techniques"))
-
-    def build_ui_page_4(self):
-        self.ka_view_model = models.KataTableViewModel(self)
-        self.ki_view_model = models.KihoTableViewModel(self)
-
-        # enable sorting through a proxy model
-        self.ka_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.ka_sort_model.setDynamicSortFilter(True)
-        self.ka_sort_model.setSourceModel(self.ka_view_model)
-
-        self.ki_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.ki_sort_model.setDynamicSortFilter(True)
-        self.ki_sort_model.setSourceModel(self.ki_view_model)
-
-        frame_ = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(frame_)
-
-        self.kata_view = self._build_kata_frame(self.ka_sort_model, vbox)
-        self.kiho_view = self._build_kiho_frame(self.ki_sort_model, vbox)
-
-        self.tabs.addTab(frame_, self.tr("Powers"))
-
-    def build_ui_page_5(self):
-        mfr = QtWidgets.QFrame(self)
-        vbox = QtWidgets.QVBoxLayout(mfr)
-
-        # advantages/disadvantage vertical toolbar
-        def _make_vertical_tb(tag, has_edit, has_remove):
-            vtb = widgets.VerticalToolBar(self)
-            vtb.addStretch()
-
-            cb_buy = (self.sink2.act_buy_merit if tag == 'merit'
-                      else self.sink2.act_buy_flaw)
-            cb_edit = (self.sink2.act_edit_merit if tag == 'merit'
-                       else self.sink2.act_edit_flaw)
-            cb_remove = (self.sink2.act_del_merit if tag == 'merit'
-                         else self.sink2.act_del_flaw)
-
-            vtb.addButton(QtGui.QIcon(get_icon_path('buy', (16, 16))),
-                          self.tr("Add Perk"), cb_buy)
-
-            if has_edit:
-                vtb.addButton(QtGui.QIcon(get_icon_path('edit', (16, 16))),
-                              self.tr("Edit Perk"), cb_edit)
-
-            if has_remove:
-                vtb.addButton(QtGui.QIcon(get_icon_path('minus', (16, 16))),
-                              self.tr("Remove Perk"), cb_remove)
-
-            vtb.addStretch()
-            return vtb
-
-        self.merits_view_model = models.PerkViewModel('merit')
-        self.flaws_view_model = models.PerkViewModel('flaws')
-
-        self.merits_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.merits_sort_model.setDynamicSortFilter(True)
-        self.merits_sort_model.setSourceModel(self.merits_view_model)
-
-        self.flaws_sort_model = models.ColorFriendlySortProxyModel(self)
-        self.flaws_sort_model.setDynamicSortFilter(True)
-        self.flaws_sort_model.setSourceModel(self.flaws_view_model)
-
-        merit_view = QtWidgets.QTableView(self)
-        merit_view.setSortingEnabled(True)
-        merit_view.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.Interactive)
-        merit_view.horizontalHeader().setStretchLastSection(True)
-        merit_view.horizontalHeader().setCascadingSectionResizes(True)
-        merit_view.doubleClicked.connect(self.sink2.act_view_merit)
-        merit_view.setModel(self.merits_sort_model)
-        merit_vtb = _make_vertical_tb('merit', True, True)
-        fr_ = QtWidgets.QFrame(self)
-        hb_ = QtWidgets.QHBoxLayout(fr_)
-        hb_.setContentsMargins(3, 3, 3, 3)
-        hb_.addWidget(merit_vtb)
-        hb_.addWidget(merit_view)
-        vbox.addWidget(new_item_groupbox(self.tr("Advantages"), fr_))
-
-        flaw_view = QtWidgets.QTableView(self)
-        flaw_view.setSortingEnabled(True)
-        flaw_view.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.Interactive)
-        flaw_view.horizontalHeader().setStretchLastSection(True)
-        flaw_view.horizontalHeader().setCascadingSectionResizes(True)
-        flaw_view.doubleClicked.connect(self.sink2.act_view_flaw)
-        flaw_view.setModel(self.flaws_sort_model)
-        flaw_vtb = _make_vertical_tb('flaw', True, True)
-        fr_ = QtWidgets.QFrame(self)
-        hb_ = QtWidgets.QHBoxLayout(fr_)
-        hb_.setContentsMargins(3, 3, 3, 3)
-        hb_.addWidget(flaw_vtb)
-        hb_.addWidget(flaw_view)
-        vbox.addWidget(new_item_groupbox(self.tr("Disadvantages"), fr_))
-
-        self.merit_view = merit_view
-        self.flaw_view = flaw_view
-        self.table_views.append(self.merit_view)
-        self.table_views.append(self.flaw_view)
-
-        self.tabs.addTab(mfr, self.tr("Perks"))
 
     def build_ui_page_7(self):
         self.melee_view_model = models.WeaponTableViewModel('melee', self)
@@ -845,92 +480,6 @@ class L5RMain(AboutTabMixin, AdvancementsTabMixin, AdvanceMixin, AdviseMixin,
         self.bt_edit_school.clicked.connect(self.sink4.on_edit_first_school)
 
         self.bt_set_exp_points.clicked.connect(self.sink1.on_set_exp_limit)
-
-    def do_buy_kata(self, kata):
-        """attempt to buy a new kata"""
-        if self.buy_kata(kata) == CMErrors.NOT_ENOUGH_XP:
-            self.not_enough_xp_advise(self)
-
-    def do_buy_kiho(self, kiho):
-        """attempt to buy a new kiho"""
-        if self.buy_kiho(kiho) == CMErrors.NOT_ENOUGH_XP:
-            self.not_enough_xp_advise(self)
-
-    def on_buy_skill_rank(self):
-        # get selected skill
-        sm_ = self.skill_table_view.selectionModel()
-        if sm_.hasSelection():
-            model_ = self.skill_table_view.model()
-            skill_id = model_.data(sm_.currentIndex(), QtCore.Qt.UserRole)
-
-            err_ = self.buy_next_skill_rank(skill_id)
-            if err_ != CMErrors.NO_ERROR:
-                if err_ == CMErrors.NOT_ENOUGH_XP:
-                    self.not_enough_xp_advise(self)
-                return
-
-            idx = None
-            for i in range(0, self.skill_table_view.model().rowCount()):
-                idx = self.skill_table_view.model().index(i, 0)
-                if model_.data(idx, QtCore.Qt.UserRole) == skill_id:
-                    break
-            if idx.isValid():
-                sm_.setCurrentIndex(idx, (QtCore.QItemSelectionModel.Select |
-                                          QtCore.QItemSelectionModel.Rows))
-
-    def act_memo_spell(self):
-        # get selected spell
-        sm_ = self.spell_table_view.selectionModel()
-        if sm_.hasSelection():
-            model_ = self.spell_table_view.model()
-            spell_itm = model_.data(sm_.currentIndex(), QtCore.Qt.UserRole)
-
-            err_ = CMErrors.NO_ERROR
-            if spell_itm.memo:
-                self.remove_advancement_item(spell_itm.adv)
-            else:
-                err_ = self.memo_spell(spell_itm.spell_id)
-
-            if err_ != CMErrors.NO_ERROR:
-                if err_ == CMErrors.NOT_ENOUGH_XP:
-                    self.not_enough_xp_advise(self)
-                return
-
-            idx = None
-            for i in range(0, self.spell_table_view.model().rowCount()):
-                idx = self.spell_table_view.model().index(i, 0)
-                if model_.data(idx, QtCore.Qt.UserRole).spell_id == spell_itm.spell_id:
-                    break
-            if idx.isValid():
-                sm_.setCurrentIndex(idx, (QtCore.QItemSelectionModel.Select |
-                                          QtCore.QItemSelectionModel.Rows))
-
-    def act_buy_spell(self):
-        dlg = dialogs.SpellAdvDialog(self.pc, 'freeform', self)
-        dlg.setWindowTitle(self.tr('Add New Spell'))
-        dlg.set_header_text(
-            self.tr("<center><h2>Select the spell to learn</h2></center>"))
-        if dlg.exec_() == QtWidgets.QDialog.Accepted:
-            self.update_from_model()
-
-    def act_del_spell(self):
-        # get selected spell
-        sm_ = self.spell_table_view.selectionModel()
-        if sm_.hasSelection():
-            model_ = self.spell_table_view.model()
-            spell_itm = model_.data(sm_.currentIndex(), QtCore.Qt.UserRole)
-
-            if spell_itm.memo:
-                return
-            self.remove_spell(spell_itm.spell_id)
-
-    def on_spell_selected(self, current, previous):
-        # get selected spell
-        model_ = self.spell_table_view.model()
-        spell_itm = model_.data(current, QtCore.Qt.UserRole)
-
-        # toggle remove
-        self.del_spell_bt.setEnabled(not spell_itm.memo)
 
     def update_from_model(self):
